Reviews/requirements.txt
flask==2.0.3
imdbpy==2021.4.18
redis==4.2.2
waiting
